Added PrivacyPolicy.tsx: add common component & restruct bottomsheet
