+++ src/components/Modal/ModalSubComponents.tsx
@@ -3,6 +3,7 @@ import {
   ModalContentProps,
   ModalTitleProps,
 } from '@/interfaces';
+import PrivacyPolicy from '../PivacyPolicy';
 
 export function ModalTitle({ children }: ModalTitleProps) {
   return (
@@ -44,12 +45,6 @@ export function ModalButton({
   );
 }
 
-// 개인정보보호동의 페이지 추가 예정
-interface ModalPrivacyPolicyProps {}
 export function ModalPrivacyPolicy() {
-  return (
-    <span className="font-primary-lightblue inline-block border-b-2 border-[#5F9FC0] font-sfpro text-2xs font-bold opacity-20">
-      개인정보 수집 및 이용 정책 {'>'}
-    </span>
-  );
+  return <PrivacyPolicy />;
 }
