+++ src/interfaces/index.ts
@@ -5,3 +5,4 @@ export * from './icon.interface';
 export * from './frame.interface';
 export * from './route.interface';
 export * from './previous-page.interface';
+export * from './policy.interface';
