+++ src/interfaces/policy.interface.ts
@@ -1,0 +1,4 @@
+export interface PrivacySectionProps {
+  title: string;
+  content: string;
+}
